fix: [#xxxxx] bump fastramqpi version (and dependencies)
A few dependencies were outdated, so they had to be bumped as well.
Bump of FastAPI introduced breaking changes to Testclient
and bump-testclient(https://github.com/Kludex/bump-testclient#transformations)
has been used to fix this issue.

Also, it seems that adguid_sync assumes that "CLIENT_ID" defaults to "dipex",
which is not the case anymore, which made the tests fail

I had to add pytest-asyncio to fix test coverage in pipeline